fixed the difference between lines_db and lines_cmt..
this is quite strange..

previously. say we have just finished reading one file.
lines_db first tries to get the revisions associated with the
first line of the next file.. then tries to get the filename of that
next file.

this can fail when the first line of the next file is not associated
with any cvs comment.. (cause readNextLine() to return false)

by first get the filename and then get the revisions associated with
the first line solves the problem. (this is because even if no
revisions are associated with the first line, readNextLine() still
return true)

now this fixes the problem in korganizer, so meaning the first line of
some file is not associated with any cvs comment.
when I tried to find which file is the culprit, there isn't one..


git-svn-id: file:///home/xapian-svn/xapian/trunk@3519 6b2f1b92-a4f3-0310-943f-de3cbe41748b