Alter yesterdays changes so that raw_search WILL process paging directives at present and at least also get TOPDOC properly.
To do this I introduced want_paging as well as want_rset so that the various parts of the program that realise they will want certain processing merly indicate it and it will happen later on, instead of having foul complex conditions merely so we can embed the block of code in the middle of the same complex condition.

Open to suggestions on this, but it's a pattern I've observed that it's much better to take a good look at all the CGI params first and work out what needs doing, and then do it; than to do it as you go along.

Of course in this case part of "do it" is working out what to do next, but...


git-svn-id: file:///home/xapian-svn/xapian/trunk@3772 6b2f1b92-a4f3-0310-943f-de3cbe41748b