debian/copyright: Add Java bindings copyrights and licences.
debian/rules: Run "make check" on each of the built bindings.  It's a very
lightweight test but will at least ensure that the basics work.  It also
serves to check the core library is working - xapian-core's own testsuite
is probably too slow to justify devoting autobuilder time to.

debian/control.in,debian/control-php.in: Remove "designed to be" from the
package description at Wookey's suggestion.


git-svn-id: file:///home/xapian-svn/xapian/trunk@7124 6b2f1b92-a4f3-0310-943f-de3cbe41748b