feat(snippet): Added code to pass query from omega
Since it is very important to have query in xapian-core to generate snippet which is very relevant to the query. Snippet to have query component with relevance model need access to the query.

Added code to send query from omeaga by setting query using snipper.set_setquery which needs to set a string as a query.

P.S: This changes  add one more function in snipper api.