feat(snipper):setup calculate_cosine_similarity fn
Setup what will be the input to calculate cosine similarity function
what would be the sane option. initially was thinking we had strings for
both of them, but they weren't available.
So now will use the vector docterm as it is available.
Converted queryterms also to similar vector to do similar operations.
Else one as vector and one as string will require double effort.

*added defination and dummy implementation for cosine_implementation
*changed queryterms to vector similar to docterms with postitional information.