feat(snipper):Getting queryterms using queryparser
Earlier we were using term_generator to generate the terms, now we are using
query parser.
Still problems:
When the query comes it is appended with a uppercase Z, but when we parse it gets converted in Z . so we have to use lowercase z . using uppercase z doesn't work.
Moreover since omega is sending the query to the function . if i pass query as location.
what queryfunction gets is "ZSBlocat\tZlocat"
after parsing this what we get is zsblocat
so difficult to avoid z and need to have lowercase z. if we can pass the untouched query. i guess i took the parsed query from omega.. need to replace it with the actual query string
one other solution is to split it on "\t" but that will make it omega specific. even now it is .

omega passed ZSBlocat\tZlocat\tZSBpublish\tZpublish for query location publish