Suppress doxygen/Qt warning about unsupported OS X version
This used to be a #error, which we silenced with a sed regex (since
doxygen doesn't actually use the non-portable parts of the headers).

It is now just a #warning, but gets reported for almost every source
file in doxygen, which is annoyingly noisy, so adjust the regexp to
still comment it out.

(cherry picked from commit 0bbaa77b46fe6c3845ffa4e2d9d674b4424f344b)