Get some MatchSpy functionality running again
This commit brings back support for the core MatchSpy implementor
ValueCountMatchSpy, but the perl extension of PerlMatchSpy is still not working
properly.

Something is broken and the stuff in PerlMatchSpy_perl is never called by
anything.

I need to get a better understanding for how it works before I can make
extensions work.

On a related note, I cannot get the PerlStopper stuff to work properly either,
and it doesn't seem to be tested very well..