Avoid potential DB corruption With full-compaction
The item size is stored in 2 bytes, but the top bit is used to store a
flag indicating that the item is compressed with zlib.

By default items are < 1/4 of the block size, and the maximum block size
is 64K, so all is well.  However, if a database is compressed with full
compaction and a 64K block size then items > 32KB are possible, but the
size will be truncated.

This change clamps the max item size to be <= 32767 bytes.