Avoid potential DB corruption With full-compaction
The item size is stored in 2 bytes, but the top bit is used to store a
flag indicating that the item is compressed with zlib.

By default items are < 1/4 of the block size, and the maximum block size
is 64K, so all is well.  However, if a database is compressed with full
compaction and a 64K block size then items > 32KB are possible, but the
size will be truncated.

This change clamps the max item size to be <= 32767 bytes.

(cross-ported from equivalent fix to glass backend in commit
14c098c8ff00962978d7632e52dd5c3a57e73d74)

(cherry picked from commit 130243e0b253ca7c07ed7a8da6d28c8bda786195)