Overhaul database compaction API
Add a Xapian::Database::compact() method, with the Database object
serving as the source database(s).  Xapian::Compactor is now just
a functor if you want to control progress reporting and/or the merging
of user metadata.

Reimplement the existing API in terms of the new one, but mark it as
deprecated.