Overhaul the "per-site" terms
'H' prefix is hostname as before, except that if the term would be > 240 bytes
the end is hashed is the same way 'U' prefix terms are.

'P' terms are now added for every directory level, not just the start URL's
path.

And a new 'J' prefix term is added with the start URL (less any trailing '/'),
which means all files indexed from a particular "site" are now indexed by one
term. [see #376]