Try to avoid the shell for --filter commands
Just check for a set of "bad characters" being present, but implement
support for setting environment variables before the command, so
that we don't have to put '=' in that set, as it's relatively common
in command line arguments - e.g. 'abiword --to=txt --to-name=fd://1'.
Setting environment variables is probably less common, but probably
used, e.g. 'LANG=C some-filter-command'.