Document actual min GCC version as 4.7
GCC 4.7 is the minimum version which AX_CXX_COMPILE_STDCXX_11 will select.
It also seems to be available sufficiently widely that requiring it doesn't
restrict portability.

Looking at C++11 features for which support was added in 4.7, for example
there's non-static data member initialisers, which can make for clearer and
more robust code in classes with multiple constructors.