Reimplement MSet and MSetIterator
MSetIterator internally now counts down to the end of the MSet, so the
end test is now against 0, rather than against mset.size().

More of the trivial methods are now inlined, which reduces the number
of relocations needed to load the library, and should give faster code
which is a very similar size to before.