Remove DatabaseReplica methods which we don't need
For internal use we don't need reference counted internals, copy
constructor, assignment operator or the close() method.  The close()
method was used, but only in the testsuite, and those uses can be
easily eliminated by adding an extra block scope.

These method are hang-overs from it being written to be a public API,
but the current plan to make it one decouples the server and the
changesets more so will look different to this.