Default to only building shared libraries.
Building both shared and static means having to compile the files which
make up the library twice on most platforms.  Shared libraries are the
better option for most users, and if anyone really wants static libraries,
they can configure with --enable-static (or --enable-static=xapian-core if
configuring a combined tree with the bindings).