Wrap public member variables of PostingSource
These are now wrapped by getters and/or setters, depending on whether they
should be readable, writable or both.  The old public members are now
flagged as deprecated, but still work for now.

This allows full wrapping of posting source for the bindings, and there's
a feature test of this for python and python3.  Fixes #499, reported by
Joost Cassee.