[glass] Delay stepping back to first component
If a cursor is moved to an entry which doesn't exist, we need to step
back to the first component of previous entry before we can read its
tag.  However we often don't actually read its tag (e.g. if we only
wanted the key), so make this stepping back lazy so we can avoid
doing it when we don't want to read the tag.