Refactor glass backend key comparison
Instead of comparing Key objects, compare BItem and LeafItem objects by
their keys, which is mainly an enabler for future changes.

Instead of operator== and operator<, provide a compare() function
which returns negative, 0 or positive, like strcmp(), memcmp() and
std::string::compare().  This allows us to avoid a final compare to
check for equality in GlassTable::find_in_leaf(), and to terminate
early if the binary chop hits the exact entry.