[glass] Elide component_of counter in first leaf
Store a "first component" flag in the flag bits of item - when set we
know this is the first component and so we don't need to store X2.  This
directly saves 2 bytes per entry in the Btree, plus additional space due
to fewer blocks and fewer levels being needed as a result.  This
particularly helps the position table, which has a lot of entries, many
of them very small.  The saving would be expected to be a little less
than the saving from the change which shaved 2 bytes of every Btree item
in 1.3.4, since we're only shaving 2 bytes off every *entry* (an entry
with a large tag may be split into multiple items), and a simple test
suggests a saving of several percent in total DB size, which fits that.

This reduces the maximum component size to 8191, which affects tables
with a 64KB blocksize in normal use and tables with >= 16KB blocksize
with full compaction.