Ensure we can actually link with sys_errlist &c.
Previously we just checked that the symbol was declared, but
the Android NDK declares it without including an implementation
in the library, so compilation succeeds but linking fails.

As before, we care about both _errlist and _nerr, and we check
for both sys_errlist and _sys_errlist variants.

(cherry picked from commit 96d31c85c5b2d8d39f18afbe303d540479c6faf0)