Run pod2html when building from a source tarball
In 1.3.5 we were failing to ship the HTML output, as reported by
Anthony Basile.

The POD to HTML conversion functionality comes as standard with Perl,
so instead of shipping the generated docs, do the conversion when
building from a tarball.

For extra robustness, we now do the conversion using the bundled
Pod::Html module from the Perl interpreter found by configure, rather
than assuming pod2html is on PATH, and the right one (this matters
more now that it's not just people building from git who build these
docs).