Fold queryparsertest into apitest
Now it uses the backend manager machinery in the testharness, so we
don't have to hard-code use of inmemory and chert backends, but can
instead be tested with all backends which support the required features.

This also fixes test failures when both chert and glass are disabled,
due to trying to run spelling tests with the inmemory backend.