Simpler handling of platforms with bad timer_create()
AIX and GNU Hurd both have timer_create() but it always seems to fail
(on Hurd this is because there's a dummy implementation in glibc
which fails with ENOSYS).

It's not useful to compile in code which tries to use timer_create(),
on such platforms, so we want to avoid defining HAVE_TIMER_CREATE when
it's not a useful implementation.

Probing for this properly seems to require a run-time test, which is
unhelpful when cross-compiling, so for now just test against a blacklist
of platforms we know are bad.