Check _POSIX_MONOTONIC_CLOCK
If it's not defined, use CLOCK_REALTIME instead of CLOCK_MONOTONIC.

The existing hard-coded platform checks still seem to be needed, as
CLOCK_MONOTONIC is available for some functions, but doesn't work
with timer_create() for one reason or another.  But this should
avoid failures on platforms without any monotonic clock support.