Replace common/unaligned.h with common/wordaccess.h
New header implements both aligned and unaligned access to words (the
revision number at the start of each block is at fixed offset 0, so we
can rely on it being aligned).

Unaligned access functions now use memcpy() and (on little-endian
platforms) a byte-swap, which is significantly faster than the old
functions.

The new word writing functions assert if the value would be truncated.

New functions only take a pointer, rather than a pointer and offset, so
validating the offset in cases where it isn't constant (or bounded)
is now a job for the caller.  The caller has more idea what offsets are
valid, so it makes more sense for it to check.