termgenerator_internal.cc: add termcover relevance for snippets
The default snippet generation algorithm defines the relevance of a snippet
as the total of its word weights and prefers occurrences of query terms
over non-query terms. However, it does not aim at covering as many query
terms as possible. This patch introduces a relevance scheme that seeks
to maximise query term coverage within a snippet.

For example, given the (non-sensical, but hopefully illustrative) document

    quack quack moooh quack quack

and the query

    Query((Zquack@1 AND Zmoooh@2))

the following snippet of maximum length 15 is currently generated

    <b>quack</b> <b>quack</b>...

or alternatively for boolean weights

    ...<b>quack</b> <b>quack</b>

The termcover relevance strategy in this patch allows terms to contribute
their relevance only once to a snippet. This forces MSet to find a snippet
covering as many query terms as possible.

For the example document, it generates

    ...<b>moooh</b> <b>quack</b>...

The termcover option is ignored for queries containing phrases and
wildcards, and is only used if the SNIPPET_TERMCOVER flag is set.