Fix Python3 bindings to work out of the box
The SWIG-generated code to load the compiled module changed recently,
and was no longer working once we renamed xapian.py to
xapian/__init__.py.

There's no way to configure SWIG to produce what we need currently, so
for now we run a "fixup" script over the generated wrapper which
replaces the complex import code with a single statement which does
what we want.  It also strips out various unneeded code, mostly to do
with handling Python 2.x (which we don't need since we have an entirely
separate set of bindings for that).

Reported by James Aylett in #731.

The way we were importing the bindings to generate the .pyc and .pyo
files was masking this problem, so rework that as suggested by James.

(cherry picked from commit e32b88fc04c49b35c207f777d9a04b141bbe615c)