Set CLOEXEC on the lock file fd and pipe
If we're using OFD fcntl locks (as we will on modern Linux), then
CLOEXEC is important as the lock fd is held by the main process
and the lock is preserved over fork().

And if we're using traditional fcntl locks with a pipe to a child
process, we want the parent end of the pipe to be closed if the
parent calls fork() and exec().