DLH,DPH: Eliminate lower_bound
The idea of dealing with an algorithm which allows a term to have a
negative weight contribution by calculating a lower_bound for each term
and adjusting both sumpart and maxpart by it is fundamentally flawed -
the total offset it adds to a document depends on what combination of
terms it matches, and so in general isn't the same for every matching
document.

So get rid of lower_bound entirely, and instead clamp each term's
weight contribution to be >= 0.