Eliminate unnecessary ConstDatabaseWrapper class
This class aims to prevent the user calling non-const methods on the
Xapian::Database object passed to Xapian::PostingSource, but this
object is passed as a const reference, so non-const methods can't
be called on it anyway - attempts to do so fail to compile!

Eliminating this fixes at least one obscure bug -
get_doclength_lower_bound() and get_doclength_upper_bound() were
missing from ConstDatabaseWrapper, so the fallback implementations
in the base class were being used, which return much less good
bounds for backends which implement tracking of these bounds.
New testcase postingsourcebounds1 serves as a regression test for
this bug.

See https://trac.xapian.org/ticket/332 for the dicussion which led
to ConstDatabaseWrapper being added originally.