$filters now includes DATEVALUE
This means we'll force the first page when reloading or changing page
starting from existing URLs upon upgrade to 1.4.1, but the exact same
existing URL could be for a search without the date filter where we want
to force the first page, so there's an inherent ambiguity there.
Forcing first page in this case seems the least problematic side-effect.