Replace uses of useconds_t by unsigned
In 1.3.x we deprecated Xapian::timeout (a typedef for unsigned) and
started to use useconds_t instead.

However, useconds_t is now obsolete and anyway was intended to
represent a time in microseconds (confusing when Xapian's timeouts are
in milliseconds).

The Linux usleep man page notes: "Programs will be more portable if
they never mention this type explicitly."