Remove unused BOOST_FACTOR machinery from snippet code
BOOST_FACTOR has always been hard-wired to 0, so no boost is applied.

Some way to factor in that context to a relevant term or phrase tends
to enhance the relevance would be useful, but I'm not convinced this is
the best approach for achieving that.

(cherry picked from commit 82ca173f6575eda13f1ba50a63854a1e513f9648)