Rename misleading internal sequential_mode references
These methods and variables are actually tracking whether the table is
sequential (i.e. whether all leaf blocks are in key order in the file)
which is not the same as "sequential mode", which is a concept whereby
we try to fill blocks when a sequence of adjacent entries are added
in order.