GlassVersion::create() no longer calls write() and sync()
If we're compacting, we then update the version file and write again,
so this write is unhelpful there.  When creating an empty database it
is better to create the version file, write it, create the tables, and
only then sync it, as that gives the write time to be processed before
the sync.