Discount relevance of repeated terms in snippets
A snippet which matches more of the terms in the query is likely to be
better than one which just matches the same term or terms multiple
times, but a repeated term is still interesting, just less so than the
first occurrence, and less and less the more times it occurs.  So we now
halve the weight for each repeat, giving an exponential decay in the
relevance of repeated terms.

Issue raised by Robert Stepanek in
https://github.com/xapian/xapian/pull/117

(cherry picked from commit 4e5698355532fb050e88a8e1bb95c8637e17759c)