Don't round matches_estimated in serialise_mset
MSet::get_matches_estimated() rounds the estimate lazily, but
MSetPostList::get_termfreq_est() returns the estimate, and the raw
estimate is better for that.

It is also cleaner that a round-trip through serialisation gives you
an object which is as close to the original as possible.

(cherry picked from commit ed6c94bacf12dab9843da8af51a1982b3e1a2875)