snippets: Adjust max_tw up by 1/64th
This means that a term with zero termweight (e.g. one which doesn't
exist in the database, or which the current weighting scheme gives
zero weight to as happens with common terms with some weighting
schemes) will be preferred to a second occurrence of the query
term with the highest termweight.

Split snippet_termcover1 up - the BoolWeight cases are now in
snippet_termcover2.