Remove feature_internal and Make Feature subclasses independent of FeatureList::Internal
Removed feature_internal , replaced it with featurelist_internal which is present in xapian-letor/api as it is being used only by FeatureList. Since, after computing, FeatureList would populate the required statistics in each Feature Subclass there was no need for Feature sub classes to refer back to FeatureList. Also Made Changes to the Makefile so as to consider the fact that feature_internal has been removed.