Add Point, Cluster and ClusterSet classes to Cluster API
Point class is useful to provide an interface to use the document
vectors which are initialized from Documents
Cluster class stores an individual cluster i.e. A set of documents
and methods to interact with the cluster
ClusterSet class will store all the Clusters that have been created
by the Clusterer.