Add MSet methods to re rank and replace weights
Two new methods, replace_weights and sort_by_relevance, have been
added to MSet which can be used to replace weights of all
the documents in a MSet and then sort the MSet on the basis of
these new weights.

The main idea behind adding this functionality to the API was to
help xapian-letor replace weights of all the documents with the
new weights returned by the ranker and then re-rank the MSet on the
basis of these new weights. This will help xapian-letor integrate
more smoothly into xapian. replace_weights can be used only to replace
weights of the entire MSet. sort_by_relevance can then be called to
sort the MSet on the basis of weights first then the docids.
Have added tests for the new functionality.