ExpandWeight: clean up termfreq code
The only functional change should be that an approximated termfreq is
now rounded to the nearest integer at the end of the calculation -
previously it was always rounded down.

The other interesting change is removing a pointless check to ensure
that the approximated termfreq is >= rtermfreq, which must always be
true since we only scale *up* the termfreq and it's >= rtermfreq to
start with.