stemtest: Handle compressed test data
The arabic test data is compressed, because it's too large for
github to handle.  Currently this means stemtest ignores it -
this patch extends it to handle gzip compressed test data.

It's a work in progress still - we need to sort out disabling
this when zlib isn't found (or make zlib a hard requirement)
and ideally split out ZLIB_LIBS from XAPIAN_LIBS so we can
link stemtest with just zlib, and not any other libraries which
the library needs.