Reimplement Document::Internal
TermIterator from a Document now keeps a reference to the document
(new regression testcase orphaneddoctermitor1).

Term positions are now stored in a custom vector implementation which
uses no external storage when small ("Small Vector Optimisation") and
eliminates a lot of copying that was previously happening (thanks to
a combination of C++11 move semantics and copy-on-write).