git-tag-release: Make deprecation warnings non-fatal
The code on this branch still uses std::auto_ptr, which was
deprecated in C++11 and is due to be removed from the C++17
standard.  Current compilers give deprecation warnings about
it, so just make these non-fatal.

Switching all uses to std::unique_ptr doesn't just work, and it
doesn't seem worth expending a lot of effort on this for a branch
that we're about to retire.