Reimplement Xapian::MSet
get_termfreq() and get_termweight() now just return 0 rather than
throwing exceptions when called on a default-constructed MSet, which
is makes it easier to use such an MSet as a placeholder for an empty
search.

get_termweight() also now returns zero when called on a term which
wasn't in the query rather than throwing an exception - this seems
logical as such terms don't contribute weight to any documents in
the MSet.