Reimplement Database and WritableDatabase
WritableDatabase::remove_spelling() now works more sensibly with shards
- it tries to remove the spelling from each shard until all of freqdec
has been dealt with, and now returns any "unused" freqdec.  To support
this for the remote backend, the remote protocol major version has been
bumped.

Database and WritableDatabase are now reference counted handles like
most other API classes - previously they worked like a typedef to
vector<shard> which makes copying and assignment O(#shards).  The
main observable difference is that if you add a shard to a copy,
it's now added to the original (which is consistent with other API
classes, e.g. adding a document to an RSet).

WritableDatabase has supported sharding since 1.3.2, but this is now
better documented.

Adding a read-only shard to a WritableDatabase is now statically
caught at compile time in simple cases, and if the static check is
bypassed then InvalidOperationError is thrown when you attempt a
WritableDatabase operation on that shard.  Previously such a shard acted
as a black hole which quietly accepted and ignored changes.

Operations on a database with a single shard should have a bit less
overhead.

Spelling correction with multiple shards should be faster.