Cleaner handling of "installable" xapian-config
We want to strip out the "uninstalled" mechanisms from the version of
xapian-config we install, because otherwise build paths get included,
which makes the build non-reproducible, and also risks that behaviour
may depend on whether these paths exist on the target system.

We used to hard-code use of $(transform), which seems to be intended
to be a mechanism internal to automake.  Instead create the installable
version in a subdirectory and install is in the normal way.

This fixes "make uninstall" to remove xapian-config when a
program-prefix or -suffix is in use (e.g. there's a default -1.5 suffix
for git master currently).