Fix percent when only some shards have positions
If a shard without positions was the final one, this would lead
to under counting the total number leaf of subqueries which would
lead to incorrect positional calculations (and a division by zero
if the top level of the query was positional.

This bug was introduced in 1.4.3.

(cherry picked from commit ece489136e00f68498b08844602e6cf3eb917b60)