Handle failing commands in shell functions better
If a shell function is called in a context where set -e doesn't
apply (such as in the condition of an "if") then set -e is disabled
during the execution of code in the function too.

So for functions which return a value add " || exit $?" after any
command whose failure should cause bootstrap to also fail.