Remove internaltest testcase static_assert1
Now we have eliminated STATIC_ASSERT_UNSIGNED_TYPE and
STATIC_ASSERT_TYPE_DOMINATES, this testcase was only testing that the
standard C++11 static_assert worked, and only that it passed for some
cases where it should - testing expected failures is harder.  Overall
such testing seems better left to the testsuite of the compiler.