Make TEST_EQUAL() arguments side-effect free
TEST_EQUAL() is a macro which evaluates its arguments a second time
if the test fails, to report their values.  This isn't ideal and
really ought to be addressed, but for now fix uses where the argument
has side-effect (e.g. *i++) such that the reported value should match
the tested value.