Improve compiler visibility attribute detection
Test that using the attributes doesn't result in a warning - previously
we'd enable them even on platforms which don't support them, which
would result in a compiler warning for every file.

Probe for -fvisibility=hidden and -fbisibility-inlines-hidden
together.  It seems all compilers implement both or neither, and
it's faster to do one probe instead of two.