Fetch the doclength upper bound at most once
For a synonym, we use the doclength upper bound for the wdf upper bound,
and if both stats are wanted, we fetched it twice.  This should not be
hugely costly, but it's simple to just fetch it once if either is needed
and use the value for both.

(cherry picked from commit d3a368c5a44b5cbc31ee48c848f0b6c4c2a842f7)