Pass doclen and unique_terms down from PostListTree
Previously postlists called down the tree to the leaves to fetch
these stats to use when calculating weights.  In the flint backend
and before, the postlists stored the document length which is why we
implemented it this way, but in chert we changed that to store a single
list of document lengths (which is more update-friendly and needs less
space because we don't redundantly store each document length many
times).

Now we no longer need to support flint-style storage, it makes more
sense to pass down these stats from PostListTree, with a mechanism for
explicitly fetching them from PostListTree in special cases.