Make get_unique_terms() more efficient for glass
We approximate get_unique_terms() by the length of the termlist
(which counts boolean terms too) but clamp this to be no larger
than the document length.  Since we need to open the termlist to
get its length, it makes more sense to get the doclength from that
termlist for no extra cost rather than looking it up in the postlist
table.