Reduce overheads of PostList positional data support
Probably the majority of PostList objects aren't involved in phrases
or OP_NEAR during the match, so avoid the overhead of a position
list object when we know we won't need it.  For glass, we lazily
create it (so only have the overhead of a pointer, while for honey
we use a special subclass to eliminate the size overhead completely).

We also now avoid copying interpolative coded positional data in
order to iterate it in most cases.