Test harness refactor
The datadir is now specified at construction time, and remote,
singlefile and multi backend managers now take a sub-manager to specify
the underlying backend to use, which means we don't need to switch
behaviour based on a string name, and allows the backend-specific code
to be better compartmentalised in the corresponding manager subclass.

Singlefile now creates databases by compacting the corresponding
underlying backend database (creating it first if need be) rather than
always creating a temporary database to compact.