xapian-inspect: Commands now expect escaped key
The key specified to commands "goto" and "until" is now expected to
be escaped in the same way that keys are escaped for display.  This
makes it much simpler to interact with tables with binary keys.

(cherry picked from commit db4e9d17eeb87027bd0d8147597b629a89ebe8bf)