[honey] Add index on first character of key
This allows us to jump to the first key with a given character with
only an O(1) lookup in an array which is at most 1KB in size.

Currently used by HoneyTable::get_exact_entry(), and by HoneyCursor
methods find_exact() and find_entry_ge() (which are the cursor find
variants which are heavily used).