[honey] Reduce find_entry() and find_entry_lt() use
find_entry(string()) is the same as find_entry_ge(string()).

The previous pattern of using find_entry_lt() to put the cursor in a
state where next() advances to a particular key is a bit inefficient as
we have to read items before those we actually want; it's also a wasted
cursor seek if the first action is skip_to() instead of next().  So
instead put the cursor after_end() initially, and check for that in
next() and use find_entry_ge() for the particular key at that point.